+++ src/main/java/org/morphix/lang/JavaObjects.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -11,6 +11,8 @@
  * specific language governing permissions and limitations under the License.
  */
 package org.morphix.lang;
+
+import org.morphix.reflection.Constructors;
 
 /**
  * Java objects utility methods, contains several useful methods when working with objects. It is called JavaObjects
@@ -24,7 +26,7 @@ public final class JavaObjects {
 	 * Private constructor.
 	 */
 	private JavaObjects() {
-		// empty
+		throw Constructors.unsupportedOperationException();
 	}
 
 	/**
@@ -44,5 +46,4 @@ public final class JavaObjects {
 	public static <T> T cast(final Object o) {
 		return (T) o;
 	}
-
 }