+++ src/main/java/org/morphix/lang/function/ValueFunction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -30,9 +30,14 @@ public interface ValueFunction<T> extends Supplier<T> {
 	 * @return a value
 	 * @throws Exception any exception thrown from the function
 	 */
-	T value() throws Exception; // NOSONAR
+	T value() throws Exception; // NOSONAR we want to allow generic exceptions here
 
 	/**
+	 * Returns the result of the {@link #value()} method. Any exception thrown from the {@link #value()} method is wrapped
+	 * into an {@link IllegalStateException}.
+	 *
+	 * @return the result of the {@link #value()} method
+	 * @throws IllegalStateException if an exception occurs when getting the value
 	 * @see Supplier#get()
 	 */
 	@Override