+++ src/test/java/org/morphix/convert/function/IntPredicatesTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -10,8 +10,10 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.morphix.reflection.predicates;
+package org.morphix.convert.function;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -19,6 +21,8 @@ import java.util.function.IntPredicate;
 
 import org.junit.jupiter.api.Test;
 import org.morphix.lang.function.IntPredicates;
+import org.morphix.reflection.Constructors;
+import org.morphix.utils.Tests;
 
 /**
  * Test class for {@link IntPredicates}.
@@ -26,6 +30,13 @@ import org.morphix.lang.function.IntPredicates;
  * @author Radu Sebastian LAZIN
  */
 class IntPredicatesTest {
+
+	@Test
+	void shouldThrowExceptionWhenTryingToInstantiate() {
+		UnsupportedOperationException e = Tests.verifyDefaultConstructorThrows(IntPredicates.class);
+
+		assertThat(e.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
+	}
 
 	@Test
 	void shouldNegatePredicate() {