+++ src/test/java/org/morphix/lang/function/SuppliersTest.java
@@ -10,50 +10,39 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.morphix.lang;
+package org.morphix.lang.function;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.hamcrest.Matchers.nullValue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 
 import org.junit.jupiter.api.Test;
 import org.morphix.reflection.Constructors;
 import org.morphix.utils.Tests;
 
 /**
- * Test class for {@link JavaObjects}.
+ * Test class for {@link Suppliers}.
  *
  * @author Radu Sebastian LAZIN
  */
-class JavaObjectsTest {
-
-	@Test
-	void shouldCastToInferredType() {
-		Object o = "Bubu";
-
-		String bubu = JavaObjects.cast(o);
-
-		assertThat(bubu, notNullValue());
-		assertThat(bubu, equalTo("Bubu"));
-	}
+class SuppliersTest {
 
 	@Test
 	void shouldThrowExceptionWhenTryingToInstantiate() {
-		UnsupportedOperationException e = Tests.verifyDefaultConstructorThrows(JavaObjects.class);
+		UnsupportedOperationException e = Tests.verifyDefaultConstructorThrows(Suppliers.class);
 
 		assertThat(e.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
 	}
 
 	@Test
-	void shouldThrowClassCastExceptionWhenObjectCannotBeCastToInferredType() {
-		String x = "Cucu";
+	void shouldRunRunnableAndReturnNullWhenSupplyingNullWithRunnable() {
+		Runnable runnable = mock(Runnable.class);
 
-		ClassCastException e = assertThrows(ClassCastException.class, () -> {
-			@SuppressWarnings("unused")
-			Integer xi = JavaObjects.cast(x);
-		});
+		Object result = Suppliers.supplyNull(runnable).get();
 
-		assertThat(e, notNullValue());
+		assertThat(result, nullValue());
+		verify(runnable).run();
 	}
 }